Added function to load record

+++ main.py
@@ -362,12 +362,36 @@ def openDatabaseWindow():
     def loadDatabaseViewFrame():
         def deleteRecord():
             record_name = chosen_record.get()
+            if record_name == "Select Preset":
+                return
             c.execute("DELETE FROM Presets WHERE name=?", [record_name])
             db.commit()
             records.remove(record_name)
 
         def loadRecord():
-            ...
+            global drag
+            record_name = chosen_record.get()
+            if record_name == "Select Preset":
+                return
+            c.execute("""SELECT Motion.velocity, Motion.ele_angle, Motion.azi_angle, Motion.x, Motion.y, Motion.z, 
+                        Environments.gravity, Environments.air_density,
+                        Presets.drag, 
+                        Projectiles.mass, Projectiles.drag_coefficient, Projectiles.area
+                        FROM Motion, Environments, Presets, Projectiles 
+                        WHERE Presets.EID=Environments.EID AND Presets.PID=Projectiles.PID AND Presets.MID=Motion.MID AND 
+                        Presets.name=?""",
+                      [record_name])
+            record = c.fetchall()[0]
+            drag = bool(record[8])
+
+            for value, variable in zip(record[:7], [initial_velocity, elevation_angle, azimuth_angle, x0, y0, z0, gravity]):
+                variable.set(value=value)
+
+            if drag:
+                for value, variable in zip(record[8:], [air_density, mass, drag_coefficient, surface_area]):
+                    variable.set(value=value)
+
+            loadSimFrame()
 
         def previewRecord():
             record_name = chosen_record.get()
